Update Docker configuration and environment variables
The Docker configuration in docker-compose.yaml and the environment variables in sqlserver.env were updated. The sql_server service was commented out, the SQL Server password was changed, and the ACCEPT_EULA value was updated. Furthermore, the dotnet variant in the Dockerfile under 'Dev env' was corrected and a new devcontainer.json file was added for further customization.